SCMlite/Backend/requirements.txt
fastapi
bcrypt
pyjwt
pymongo
uvicorn
passlib
requests